fix clickhouse import to only be bundled in server (#521)
