+++ frontend/package.json
@@ -23,7 +23,7 @@
     "@ai-sdk/openai": "^1.3.16",
     "@ai-sdk/react": "^1.2.9",
     "@aws-sdk/client-s3": "^3.787.0",
-    "@clickhouse/client": "1.11.0",
+    "@clickhouse/client": "^1.11.1",
     "@codemirror/lang-html": "^6.4.9",
     "@codemirror/lang-json": "^6.0.1",
     "@codemirror/lang-python": "^6.1.7",
